Conflicts:
	src/java/negocio/AdministrarMicrocurriculo.java

+++ src/java/dao/EncabezadoJpaController.java
@@ -5,17 +5,20 @@
  */
 package dao;
 
+import dao.exceptions.IllegalOrphanException;
 import dao.exceptions.NonexistentEntityException;
 import dao.exceptions.PreexistingEntityException;
 import dto.Encabezado;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import dto.EncabezadoTabla;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -33,11 +36,29 @@ public class EncabezadoJpaController implements Serializable {
     }
 
     public void create(Encabezado encabezado) throws PreexistingEntityException, Exception {
+        if (encabezado.getEncabezadoTablaList() == null) {
+            encabezado.setEncabezadoTablaList(new ArrayList<EncabezadoTabla>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<EncabezadoTabla> attachedEncabezadoTablaList = new ArrayList<EncabezadoTabla>();
+            for (EncabezadoTabla encabezadoTablaListEncabezadoTablaToAttach : encabezado.getEncabezadoTablaList()) {
+                encabezadoTablaListEncabezadoTablaToAttach = em.getReference(encabezadoTablaListEncabezadoTablaToAttach.getClass(), encabezadoTablaListEncabezadoTablaToAttach.getId());
+                attachedEncabezadoTablaList.add(encabezadoTablaListEncabezadoTablaToAttach);
+            }
+            encabezado.setEncabezadoTablaList(attachedEncabezadoTablaList);
             em.persist(encabezado);
+            for (EncabezadoTabla encabezadoTablaListEncabezadoTabla : encabezado.getEncabezadoTablaList()) {
+                Encabezado oldEncabezadoIdOfEncabezadoTablaListEncabezadoTabla = encabezadoTablaListEncabezadoTabla.getEncabezadoId();
+                encabezadoTablaListEncabezadoTabla.setEncabezadoId(encabezado);
+                encabezadoTablaListEncabezadoTabla = em.merge(encabezadoTablaListEncabezadoTabla);
+                if (oldEncabezadoIdOfEncabezadoTablaListEncabezadoTabla != null) {
+                    oldEncabezadoIdOfEncabezadoTablaListEncabezadoTabla.getEncabezadoTablaList().remove(encabezadoTablaListEncabezadoTabla);
+                    oldEncabezadoIdOfEncabezadoTablaListEncabezadoTabla = em.merge(oldEncabezadoIdOfEncabezadoTablaListEncabezadoTabla);
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findEncabezado(encabezado.getId()) != null) {
@@ -51,12 +72,45 @@ public class EncabezadoJpaController implements Serializable {
         }
     }
 
-    public void edit(Encabezado encabezado) throws NonexistentEntityException, Exception {
+    public void edit(Encabezado encabezado) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Encabezado persistentEncabezado = em.find(Encabezado.class, encabezado.getId());
+            List<EncabezadoTabla> encabezadoTablaListOld = persistentEncabezado.getEncabezadoTablaList();
+            List<EncabezadoTabla> encabezadoTablaListNew = encabezado.getEncabezadoTablaList();
+            List<String> illegalOrphanMessages = null;
+            for (EncabezadoTabla encabezadoTablaListOldEncabezadoTabla : encabezadoTablaListOld) {
+                if (!encabezadoTablaListNew.contains(encabezadoTablaListOldEncabezadoTabla)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain EncabezadoTabla " + encabezadoTablaListOldEncabezadoTabla + " since its encabezadoId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<EncabezadoTabla> attachedEncabezadoTablaListNew = new ArrayList<EncabezadoTabla>();
+            for (EncabezadoTabla encabezadoTablaListNewEncabezadoTablaToAttach : encabezadoTablaListNew) {
+                encabezadoTablaListNewEncabezadoTablaToAttach = em.getReference(encabezadoTablaListNewEncabezadoTablaToAttach.getClass(), encabezadoTablaListNewEncabezadoTablaToAttach.getId());
+                attachedEncabezadoTablaListNew.add(encabezadoTablaListNewEncabezadoTablaToAttach);
+            }
+            encabezadoTablaListNew = attachedEncabezadoTablaListNew;
+            encabezado.setEncabezadoTablaList(encabezadoTablaListNew);
             encabezado = em.merge(encabezado);
+            for (EncabezadoTabla encabezadoTablaListNewEncabezadoTabla : encabezadoTablaListNew) {
+                if (!encabezadoTablaListOld.contains(encabezadoTablaListNewEncabezadoTabla)) {
+                    Encabezado oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla = encabezadoTablaListNewEncabezadoTabla.getEncabezadoId();
+                    encabezadoTablaListNewEncabezadoTabla.setEncabezadoId(encabezado);
+                    encabezadoTablaListNewEncabezadoTabla = em.merge(encabezadoTablaListNewEncabezadoTabla);
+                    if (oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla != null && !oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla.equals(encabezado)) {
+                        oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla.getEncabezadoTablaList().remove(encabezadoTablaListNewEncabezadoTabla);
+                        oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla = em.merge(oldEncabezadoIdOfEncabezadoTablaListNewEncabezadoTabla);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -74,7 +128,7 @@ public class EncabezadoJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -85,6 +139,17 @@ public class EncabezadoJpaController implements Serializable {
                 encabezado.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The encabezado with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<EncabezadoTabla> encabezadoTablaListOrphanCheck = encabezado.getEncabezadoTablaList();
+            for (EncabezadoTabla encabezadoTablaListOrphanCheckEncabezadoTabla : encabezadoTablaListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Encabezado (" + encabezado + ") cannot be destroyed since the EncabezadoTabla " + encabezadoTablaListOrphanCheckEncabezadoTabla + " in its encabezadoTablaList field has a non-nullable encabezadoId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(encabezado);
             em.getTransaction().commit();
